Change imports due to model refactoring
