+++ openprocurement/auctions/core/plugins/awarding/v1/models.py
@@ -3,11 +3,12 @@ from schematics.types import (
 )
 from schematics.types.compound import ModelType
 
-from openprocurement.api.models.auction_models.models import (
+from openprocurement.api.models.auction_models import (
     Organization,
     Document
 )
 from openprocurement.api.models.schematics_extender import ListType
+
 from openprocurement.auctions.core.models import (
     flashComplaint as Complaint,
     flashItem as Item,
