+++ openprocurement/auctions/core/plugins/awarding/v2/models.py
@@ -7,9 +7,9 @@ from schematics.types import StringType
 from schematics.types.serializable import serializable
 from schematics.types.compound import ModelType
 
-from openprocurement.api.utils import calculate_business_date
+from openprocurement.api.models.common import Period
 from openprocurement.api.models.schematics_extender import ListType
-from openprocurement.api.models.models import Period
+from openprocurement.api.utils import calculate_business_date
 
 from openprocurement.auctions.core.models import (
     get_auction,
@@ -19,6 +19,10 @@ from openprocurement.auctions.core.models import (
     dgfOrganization as Organization,
     Award as BaseAward
 )
+from openprocurement.auctions.core.plugins.awarding.base.constants import (
+    CONTRACT_SIGNING_TIME,
+    AWARD_PAYMENT_TIME
+)
 from openprocurement.auctions.core.plugins.awarding.v2.constants import (
     VERIFY_AUCTION_PROTOCOL_TIME
 )
@@ -26,10 +30,6 @@ from openprocurement.auctions.core.validation import (
     validate_disallow_dgfPlatformLegalDetails
 )
 
-from openprocurement.auctions.core.plugins.awarding.base.constants import (
-    CONTRACT_SIGNING_TIME,
-    AWARD_PAYMENT_TIME
-)
 
 class Award(BaseAward):
     """
