+++ openprocurement/auctions/core/plugins/awarding/v2_1/models.py
@@ -7,8 +7,8 @@ from schematics.types import StringType
 from schematics.types.serializable import serializable
 from schematics.types.compound import ModelType
 
+from openprocurement.api.models.common import Period
 from openprocurement.api.models.schematics_extender import ListType
-from openprocurement.api.models.models import Period
 
 from openprocurement.auctions.core.models import (
     get_auction,
@@ -18,15 +18,15 @@ from openprocurement.auctions.core.models import (
     dgfOrganization as Organization,
     Award as BaseAward
 )
+from openprocurement.auctions.core.plugins.awarding.base.constants import (
+    CONTRACT_SIGNING_TIME,
+    AWARD_PAYMENT_TIME
+)
 from openprocurement.auctions.core.plugins.awarding.v2_1.constants import (
     VERIFY_AUCTION_PROTOCOL_TIME
 )
 from .utils import calculate_enddate
 
-from openprocurement.auctions.core.plugins.awarding.base.constants import (
-    CONTRACT_SIGNING_TIME,
-    AWARD_PAYMENT_TIME
-)
 
 class Award(BaseAward):
     class Options:
