+++ openprocurement/auctions/core/plugins/contracting/v2/models.py
@@ -1,6 +1,6 @@
 from schematics.types.compound import ModelType
 
-from openprocurement.api.models.auction_models.models import (
+from openprocurement.api.models.auction_models import (
     ListType,
 )
