+++ openprocurement/auctions/core/tests/plugins/contracting/v3/tests/unit/contract_validation.py
@@ -1,8 +1,9 @@
 from uuid import uuid4
 from datetime import datetime
+
 from schematics.exceptions import ValidationError
 
-from openprocurement.api.models.auction_models.models import Period
+from openprocurement.api.models.common import Period
 
 from openprocurement.auctions.core.tests.base import BaseWebTest
 from openprocurement.auctions.core.plugins.contracting.v3.models import (
@@ -10,6 +11,7 @@ from openprocurement.auctions.core.plugins.contracting.v3.models import (
 )
 
 contract_data = {'awardID': uuid4().hex}
+
 
 class TestContractingV3ContractValidation(BaseWebTest):
